uint16 -> uint32

+++ protocol/contracts/beanstalk/silo/WhitelistFacet/WhitelistFacet.sol
@@ -44,7 +44,7 @@ contract WhitelistFacet is WhitelistedTokens {
     function whitelistToken(
         address token,
         bytes4 selector,
-        uint16 stalkIssuedPerBdv,
+        uint32 stalkIssuedPerBdv,
         uint32 stalkEarnedPerSeason,
         bytes4 gaugePointSelector,
         bytes4 liquidityWeightSelector,
